Added fluent validation

+++ ItemTradeApp/Features/CounterOffers/CounterOffersController.cs
@@ -35,7 +35,7 @@ public sealed class CounterOffersController(ICounterOffersService counterOffersS
         var result = await counterOffersService.UpdateCounterOfferStatusAsync(
             auth0UserId,
             counterOfferId,
-            request.StatusId,
+            (int)request.StatusId,
             ct);
 
         return result.ToActionResult();
